% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchoolDataIT-package.R
\docType{package}
\name{Get-package}
\alias{Get}
\alias{Get-package}
\title{Get: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Leonardo Cefalo \email{leonardo.cefalo@uniba.it}

}
\keyword{internal}
